requirement.txt
azure-cognitiveservices-speech
python-dotenv
pydub
openai
ipykernel
pytest
bs4
pandas
GoogleNews